services/transcribe_service/requirements.txt
fastapi[standard]>=0.116.1
openai-whisper>=20250625
pydantic>=2.11.7
google-cloud-storage>=3.3.0
tenacity>=9.1.2